# Optimise status command