New features:

- [x] Include results that are in the workspace but not the snapshot
- [x] Use modified time as a heuristic to reduce the amount of hashing performed
- [x] Calculate the sha1sum with a 4KB sliding buffer in status subcommand
- [x] Windows Symlink handle folders as well

Optional goals:

- [ ] Move some code in the library portion

The average time to set a status on a build install tree of boost version 1.80 (15 000 file) is (calculate an average of 10 runs without modification and 10 runs with modification) :

- elfshaker release -> 7.057s
- elfshaker of this pr -> 1.101s

The size of the directory is 360.8 MB :

- the largest file 988KB
- the smallest file 1KB

The average time to set a status on a build tree of llvm (2822) is (calculate an average of 10 runs without modification and 10 runs with modification) :

- elfshaker release -> 40.214 s
- elfshaker of this pr -> 0.169 s

The size of the directory is 19,6GB :

- the largest file 298MB
- the smallest file 4KB