@Lambourl 2023-10-11
I have just checked in tipi the status of this pr returns all files as changed even if none have changed.