@Lambourl 2023-10-11
all file are not listed in workspace_file_paths