@Lambourl 2023-10-14
The average time to set a status on a build tree of llvm (2822) is (calculate an average of 10 runs without modification and 10 runs with modification) :

elfshaker release -> 40.214 s
elfshaker of this pr -> 2.037s
The size of the directory is 19,6GB :

the largest file 298MG
the smallest file 4KB